sample working okay on windows
